# docs/swagger.json
{
    "schemes": [
        "http",
        "https",
        "",
        "//",
        "根据您的服务支持情况调整"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是帖子搜索服务的 API 文档。它允许搜索从 Kafka 事件中索引的帖子。",
        "title": "帖子搜索服务 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8083 // 请根据您的实际配置修改",
    "paths": {
        "/api/v1/search/hot-terms": {
            "get": {
                "description": "返回最流行或最近搜索词的列表。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "获取热门搜索词",
                "parameters": [
                    {
                        "maximum": 50,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "返回的热门搜索词数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回热门搜索词列表。",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerHotSearchTermsResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误，无法获取热门搜索词。",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/search/search": {
            "get": {
                "description": "根据关键词、分页、排序等条件搜索帖子列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "搜索帖子",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码 (从1开始)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "updated_at",
                        "description": "排序字段 (例如: updated_at, view_count, _score)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "排序顺序 (asc 或 desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功，返回匹配的帖子列表及分页信息。",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerSearchResultResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数无效，例如页码超出范围或排序字段不支持。",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误，搜索服务遇到未预期的问题。",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.EsPostDocument": {
            "type": "object",
            "properties": {
                "author_avatar": {
                    "description": "作者头像的 URL 或标识符。",
                    "type": "string"
                },
                "author_id": {
                    "description": "作者的用户 ID。",
                    "type": "string"
                },
                "author_username": {
                    "description": "作者的用户名。",
                    "type": "string"
                },
                "contact_info": {
                    "description": "联系方式",
                    "type": "string"
                },
                "content": {
                    "description": "帖子内容。",
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "highlights": {
                    "description": "新增：用于存储高亮片段的字段\n键是字段名 (如 \"title\", \"content\")，值是包含高亮HTML片段的字符串切片。\nomitempty 表示如果 Highlights 为 nil 或空 map，则在JSON序列化时忽略此字段。\n我们不在 _source 中存储这个字段，它是由 Elasticsearch 在查询时动态生成的。\n因此，不需要 `json:\"-\"` 标签来阻止它被 Elasticsearch 索引，\n但在API响应中我们希望包含它，所以使用 `json:\"highlights,omitempty\"`。",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "description": "帖子唯一标识符。使用 uint64 以兼容 ES 的 long 或 unsigned_long 类型。",
                    "type": "integer"
                },
                "images": {
                    "description": "图片列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImageEventData"
                    }
                },
                "official_tag": {
                    "description": "官方标签，直接使用导入的枚举类型（建议在 ES 中存储为整数或映射为 keyword）。",
                    "type": "integer",
                    "example": 0
                },
                "price_per_unit": {
                    "description": "每单位价格（如果适用）。",
                    "type": "number"
                },
                "status": {
                    "description": "帖子状态，直接使用导入的枚举类型（建议在 ES 中存储为整数或映射为 keyword）。",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "帖子标题。",
                    "type": "string"
                },
                "updated_at": {
                    "description": "文档在 Elasticsearch 中最后更新的时间戳。",
                    "type": "string"
                },
                "view_count": {
                    "description": "帖子浏览量。",
                    "type": "integer"
                }
            }
        },
        "models.HotSearchTerm": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "搜索词的频率计数，omitempty表示如果为0则不在JSON中显示，可选",
                    "type": "integer"
                },
                "term": {
                    "description": "搜索词本身",
                    "type": "string"
                }
            }
        },
        "models.ImageEventData": {
            "type": "object",
            "properties": {
                "display_order": {
                    "description": "Display order of the image (optional)",
                    "type": "integer"
                },
                "image_url": {
                    "description": "Publicly accessible URL of the image",
                    "type": "string"
                },
                "object_key": {
                    "description": "Key of the image in object storage (optional, if consumers need it)",
                    "type": "string"
                }
            }
        },
        "models.SearchResult": {
            "type": "object",
            "properties": {
                "hits": {
                    "description": "命中的帖子列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EsPostDocument"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "size": {
                    "description": "当前页大小",
                    "type": "integer"
                },
                "took_ms": {
                    "description": "UPRAVENO: Doba trvání dotazu v milisekundách (typ int64)",
                    "type": "integer",
                    "example": 50
                },
                "total": {
                    "description": "总命中数",
                    "type": "integer"
                }
            }
        },
        "models.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务自定义错误码。",
                    "type": "integer"
                },
                "data": {
                    "description": "错误响应中 data 字段通常为 null 或不包含有效业务数据，这里使用 interface{}。"
                },
                "message": {
                    "description": "错误的文字描述。",
                    "type": "string"
                }
            }
        },
        "models.SwaggerHotSearchTermsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务自定义状态码，例如 0 代表成功，其他值代表特定错误。",
                    "type": "integer"
                },
                "data": {
                    "description": "告诉前端哪些词是热门的。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.HotSearchTerm"
                        }
                    ]
                },
                "message": {
                    "description": "操作结果的文字描述，例如 \"搜索成功\" 或具体的错误信息。",
                    "type": "string"
                }
            }
        },
        "models.SwaggerSearchResultResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务自定义状态码，例如 0 代表成功，其他值代表特定错误。",
                    "type": "integer"
                },
                "data": {
                    "description": "具体的搜索结果数据负载。使用 omitempty 可以在 Data 为空时不显示该字段。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SearchResult"
                        }
                    ]
                },
                "message": {
                    "description": "操作结果的文字描述，例如 \"搜索成功\" 或具体的错误信息。",
                    "type": "string"
                }
            }
        }
    }
}